move metamask functions
